Revert "Feat : 전화 끊기 api, repository 생성"
This reverts commit 6b685193df14bf1caeaab558c19ec02aea4f08d4.

# Conflicts:
#	core/notification/build/kotlin/compileDebugKotlin/cacheable/last-build.bin
#	core/notification/build/kotlin/compileDebugKotlin/local-state/build-history.bin
#	core/notification/build/kspCaches/debug/classLookups/lookups.tab
#	core/notification/build/kspCaches/debug/classLookups/lookups.tab.values.at
#	core/notification/build/kspCaches/debug/classpath-structure.bin
#	core/notification/build/kspCaches/debug/symbolLookups/lookups.tab
#	core/notification/build/kspCaches/debug/symbolLookups/lookups.tab.values.at
#	core/ui/src/main/java/com/likelion/ui/component/full_screen/FullScreenCallInit.kt
#	data/src/main/java/com/likelion/data/call_for_caller/repository/CallRepositoryImpl.kt
#	feature/call/src/main/java/com/lion/call/call_for_caller/CallForCallerScreen.kt
#	remote/src/main/java/com/likelion/remote/api/CallApiService.kt

+++ app/src/main/java/com/likelion/siso/navigation/MainNavHostViewModel.kt
@@ -78,7 +78,7 @@ class MainNavHostViewModel @Inject constructor(
 
     fun callMapper(call: Call, receiverId: Long): CallModel {
         return CallModel(
-            id = call.id.toLong(),
+            id = call.id.toInt(),
             callerId = call.callerId.toLong(),
             channelName = call.agoraChannel,
             agoraToken = call.agoraToken,
